Move plugin-related utils into plugins utils module
